feat: implemented register of auctions

+++ Data/Models/Leiloes.cs
@@ -5,7 +5,7 @@ namespace CardBid.Data.Models
         public int Id { get; set; }
         public DateTime DataLimite { get; set; }
         public decimal PrecoInicial { get; set; }
-        public decimal BidFee { get; set; }
+        public double  BidFee { get; set; }
         public string Estado { get; set; }
         public int GrauDeDegradacao { get; set; }
         public string Descricao { get; set; }
@@ -15,7 +15,7 @@ namespace CardBid.Data.Models
         public string Titulo { get; set; }
         public Leiloes() { }
 
-        public Leiloes(int id, DateTime DataLimite, decimal PrecoInicial, decimal BidFee, string Estado,
+        public Leiloes(int id, DateTime DataLimite, decimal PrecoInicial, double BidFee, string Estado,
             int GrauDeDegradacao, string Descricao, int VendedorId, string Categoria,
             int? MaiorLicitacao, string Titulo)
         {
@@ -32,7 +32,7 @@ namespace CardBid.Data.Models
             this.Titulo = Titulo;
         }
 
-        public Leiloes(DateTime DataLimite, decimal PrecoInicial, decimal BidFee, string Estado, int GrauDeDegradacao,
+        public Leiloes(DateTime DataLimite, decimal PrecoInicial, double BidFee, string Estado, int GrauDeDegradacao,
             string Descricao, int VendedorId, string Categoria, string Titulo)
         {
             this.DataLimite = DataLimite;
